pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "versioneer",
    "packaging",
    "amulet-compiler-version@git+https://github.com/Amulet-Team/Amulet-Compiler-Version.git@3.0"
]
build-backend = "build_requires"
backend-path = [""]

[project]
name = "amulet-utils"
authors = [
    {name = "James Clare"},
]
description = "A C++ utility library with a python wrapper."
dynamic = ["version", "readme", "dependencies"]
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev = [
    "setuptools>=42",
    "versioneer",
    "wheel",
    "pybind11==2.13.6",
    "amulet_pybind11_extensions~=1.0",
    "pybind11_stubgen>=2.5.4",
    "black>=22.3",
    "isort",
    "autoflake",
    "mypy",
    "types-pyinstaller",
    "amulet-test-utils~=1.1",
]
numpy = [
    "numpy~=2.0",
]
pillow = [
    "pillow~=11.0",
]

[project.urls]
Homepage = "https://www.amuletmc.com"
Repository = "https://github.com/Amulet-Team/Amulet-Utils"
Issues = "https://github.com/Amulet-Team/Amulet-Utils/issues"

[tool.setuptools]
include-package-data = false

[tool.setuptools.package-data]
"*" = [
    "*Config.cmake",
    "**/*.hpp",
    "**/*.dll",
    "**/*.so",
    "**/*.dylib",
    "**/*.lib",
]

[tool.setuptools.dynamic]
readme = {file = ["README.md"], content-type = "text/markdown"}

[tool.versioneer]
VCS = "git"
style = "pep440"
versionfile_source = "src/amulet/utils/_version.py"
versionfile_build = "amulet/utils/_version.py"
tag_prefix = ""
parentdir_prefix = "amulet_utils-"
